requirements.txt
langchain
langchain_community
langchain_google_genai
qdrant-client
sentence-transformers
numpy
streamlit
python-dotenv